Initial submit of the file read/writer
Can write out an arbitrary sized bit map, with generated pixel colours as a test. This is to get the file writing working properly before moving to cleaned up code or generating images.
The commented out bits move green and red bars from the edges of the picture, and adjust the direction of the gradient fill, in order to test byte offsets and alignment properly, and so I can easily tell the orientation of the pixel bytes in the data versus how they're displayed in image applications.