Fixing an issue when indexing into the image canvas via an xy coordinate
The calculation of the pixel coordinate didn't match user expectations of the xy origin being at the top left corner of the image. Fixed the calcuation, to meet this. GetPixel operated in file arrangement of the pixels, so moved the calculation to the file writer, and swapped it to use the raw pixel buffer.